Restaurant_opencv
I've split the code into two files. The "main" file is responsible for running the code, while the "menu" file assists in the functionality of the code. The code is written in the Python language and utilizes the OpenCV and NumPy libraries. The implementation includes "Convex Hull" and "Single Responsibility Principle" (SRP). For the code to run successfully, a camera with a resolution of 1080x1920 is required. It operates with a flawlessly simple image processing system.